requirements.txt
requests
spotipy
pillow
python-dotenv